+++ lib/src/flutter_blue.dart
@@ -7,13 +7,20 @@ part of flutter_blue;
 class FlutterBlue {
   final MethodChannel _channel = const MethodChannel('$NAMESPACE/methods');
   final EventChannel _stateChannel = const EventChannel('$NAMESPACE/state');
-  final EventChannel _scanResultChannel = const EventChannel('$NAMESPACE/scanResult');
-  final EventChannel _servicesDiscoveredChannel = const EventChannel('$NAMESPACE/servicesDiscovered');
-  final EventChannel _characteristicReadChannel = const EventChannel('$NAMESPACE/characteristicRead');
-  final EventChannel _descriptorReadChannel = const EventChannel('$NAMESPACE/descriptorRead');
-  final EventChannel _characteristicNotifiedChannel = const EventChannel('$NAMESPACE/characteristicNotified');
-  final StreamController<MethodCall> _methodStreamController = new StreamController.broadcast(); // ignore: close_sinks
-  Stream<MethodCall> get _methodStream => _methodStreamController.stream; // Used internally to dispatch methods from platform.
+  final EventChannel _scanResultChannel =
+      const EventChannel('$NAMESPACE/scanResult');
+  final EventChannel _servicesDiscoveredChannel =
+      const EventChannel('$NAMESPACE/servicesDiscovered');
+  final EventChannel _characteristicReadChannel =
+      const EventChannel('$NAMESPACE/characteristicRead');
+  final EventChannel _descriptorReadChannel =
+      const EventChannel('$NAMESPACE/descriptorRead');
+  final EventChannel _characteristicNotifiedChannel =
+      const EventChannel('$NAMESPACE/characteristicNotified');
+  final StreamController<MethodCall> _methodStreamController =
+      new StreamController.broadcast(); // ignore: close_sinks
+  Stream<MethodCall> get _methodStream => _methodStreamController
+      .stream; // Used internally to dispatch methods from platform.
 
   /// Singleton boilerplate
   FlutterBlue._() {
@@ -47,6 +54,7 @@ class FlutterBlue {
   }
 
   /// Starts a scan for Bluetooth Low Energy devices
+  /// Timeout closes the stream after a specified [Duration]
   Stream<ScanResult> scan({
     ScanMode scanMode = ScanMode.lowLatency,
     List<Guid> withServices = const [],
@@ -57,10 +65,18 @@ class FlutterBlue {
       ..androidScanMode = scanMode.value
       ..serviceUuids.addAll(withServices.map((g) => g.toString()).toList());
     StreamSubscription subscription;
-    final controller = new StreamController(onCancel: () {
-      _stopScan();
-      subscription.cancel();
-    });
+    StreamController controller;
+    controller = new StreamController(
+      onListen: () {
+        if(timeout != null) {
+          new Future.delayed(timeout, () => controller.close());
+        }
+      },
+      onCancel: () {
+        _stopScan();
+        subscription.cancel();
+      },
+    );
 
     await _channel.invokeMethod('startScan', settings.writeToBuffer());
 
@@ -70,12 +86,7 @@ class FlutterBlue {
           onDone: controller.close,
         );
 
-    var stream = controller.stream;
-    if(timeout != null) {
-      stream = stream.timeout(timeout, onTimeout: (s) => controller.close());
-    }
-
-    yield* stream
+    yield* controller.stream
         .map((List<int> data) => new protos.ScanResult.fromBuffer(data))
         .map((p) => new ScanResult.fromProto(p));
   }
@@ -85,26 +96,41 @@ class FlutterBlue {
 
   /// Establishes a connection to the Bluetooth Device.
   /// Returns a stream of [BluetoothDeviceState]
-  /// When [autoConnect] is true, the connection attempt will not time out.
+  /// Timeout closes the stream after a specified [Duration]
   /// To cancel connection to device, simply cancel() the stream subscription
-  /// NOTE: iOS will never time out the connection, Android may return GATT error 133 (or others).
-  Stream<BluetoothDeviceState> connect(BluetoothDevice device, {bool autoConnect = true}) async* {
+  Stream<BluetoothDeviceState> connect(BluetoothDevice device,
+      {Duration timeout}) async* {
     var request = protos.ConnectRequest.create()
       ..remoteId = device.id.toString()
-      ..androidAutoConnect = autoConnect;
+      ..androidAutoConnect = true;
+    var connected = false;
     StreamSubscription subscription;
-    final controller = new StreamController(onCancel: () {
-      _cancelConnection(device);
-      subscription.cancel();
-    });
+    StreamController controller;
+    controller = new StreamController(
+      onListen: () {
+        if(timeout != null) {
+          new Future.delayed(timeout, () => (!connected) ? controller.close(): null);
+        }
+      },
+      onCancel: () {
+        _cancelConnection(device);
+        subscription.cancel();
+      },
+    );
 
     await _channel.invokeMethod('connect', request.writeToBuffer());
 
     subscription = device.onStateChanged().listen(
-      controller.add,
-      onError: controller.addError,
-      onDone: controller.close,
-    );
+        (data) {
+          if(data == BluetoothDeviceState.connected) {
+            print('connected!');
+            connected = true;
+          }
+          controller.add(data);
+        },
+          onError: controller.addError,
+          onDone: controller.close,
+        );
 
     yield* controller.stream;
   }
@@ -146,16 +172,17 @@ class DeviceIdentifier {
   int get hashCode => id.hashCode;
 
   @override
-  bool operator ==(other) => other is DeviceIdentifier && compareAsciiLowerCase(id, other.id) == 0;
+  bool operator ==(other) =>
+      other is DeviceIdentifier && compareAsciiLowerCase(id, other.id) == 0;
 }
 
 class ScanResult {
-  const ScanResult(
-      {this.device, this.advertisementData, this.rssi});
+  const ScanResult({this.device, this.advertisementData, this.rssi});
 
   ScanResult.fromProto(protos.ScanResult p)
       : device = new BluetoothDevice.fromProto(p.device),
-        advertisementData = new AdvertisementData.fromProto(p.advertisementData),
+        advertisementData =
+            new AdvertisementData.fromProto(p.advertisementData),
         rssi = p.rssi;
 
   final BluetoothDevice device;